fix: fix a bug
